Move SchoolsGridPage components to new files
